intelligence-per-watt/pyproject.toml
[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[project]
name = "ipw"
version = "0.1.0"
description = "Intelligence Per Watt"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "aiohttp>=3.12.15",
    "rich>=13",
    "click>=8",
    "matplotlib>=3.9",
    "grpcio>=1.62",
    "protobuf>=5",
    "pytest>=8",
    "requests>=2",
    "datasets>=2.20",
    "numpy>=1.26",
    "pyarrow>=15",
    "tqdm>=4.66",
    "hf_transfer",
    "python-dotenv",
]

[project.optional-dependencies]
ollama = ["ollama"]
vllm = ["vllm>=0.11.0"]
all = ["ollama", "vllm>=0.11.0"]

[project.scripts]
ipw = "ipw.cli:main"

[tool.uv_build]
package-data = {"ipw.telemetry" = ["bin/**"], "ipw.datasets.ipw" = ["data/**"]}
